refactor/23: Moved go code from backend/ folder to project root
