+++ .github/workflows/ci-build.yml
@@ -37,13 +37,11 @@ jobs:
 
       - name: Install dependencies
         run: |
-          cd backend
           go get -t ./...
 
       - name: Check generated code
         run: |
           go install github.com/swaggo/swag/cmd/swag@latest
-          cd backend 
           go generate
           if [[ `git status --porcelain` ]]; then
             exit 1
@@ -51,17 +49,14 @@ jobs:
           
       - name: golangci-lint
         run: |
-          cd backend
-          ../ci/install_and_run_golangci-lint.sh
+          ./ci/install_and_run_golangci-lint.sh
 
       - name: Unit tests
         run: |
-          cd backend
           go test -v -short ./...
 
       - name: Integration tests
         run: |
-          cd backend
           # -p 1 tells go not to run the tests in parallel. See the developers_guide.
           go test -v -p 1 -coverprofile=coverage.out ./...
