+++ ExasolSqlClient.go
