+++ integrationTesting/extensionForTesting/package.json
