+++ restAPI/restApi.go
@@ -1,9 +1,9 @@
 package restAPI
 
 import (
-	cont "backend/extensionController"
 	"context"
 	"database/sql"
+	cont "extension-manager/extensionController"
 	"fmt"
 	"github.com/exasol/exasol-driver-go"
 	"github.com/gin-gonic/gin"
@@ -13,7 +13,7 @@ import (
 	"sync"
 )
 
-import _ "backend/generatedApiDocs" // docs are generated by Swag CLI, you have to import it.
+import _ "extension-manager/generatedApiDocs" // docs are generated by Swag CLI, you have to import it.
 
 // RestAPI is the interface that provides the REST API server of the extension-manager.
 type RestAPI interface {
