+++ sonar-project.properties
@@ -1,8 +1,8 @@
 sonar.projectKey=com.exasol:extension-manager
 sonar.organization=exasol
-sonar.sources=backend/
-sonar.exclusions=backend/**/*_test.go,backend/integrationTesting/**,backend/generatedApiDocs/**
+sonar.sources=./
+sonar.exclusions=./**/*_test.go,integrationTesting/**,generatedApiDocs/**
 sonar.sourceEncoding=UTF-8
-sonar.go.coverage.reportPaths=backend/coverage.out
-sonar.tests=backend/
-sonar.test.inclusions=*_test.go,backend/integrationTesting/**
+sonar.go.coverage.reportPaths=coverage.out
+sonar.tests=./
+sonar.test.inclusions=*_test.go,integrationTesting/**
